chore: move barretenberg from crate root to a module
