+++ src/barretenberg/pedersen.rs
@@ -1,6 +1,6 @@
 use acvm::FieldElement;
 
-use super::{Barretenberg, Error};
+use super::{Assignments, Barretenberg, Error, FIELD_BYTES};
 
 pub(crate) trait Pedersen {
     fn compress_native(
@@ -18,8 +18,6 @@ impl Pedersen for Barretenberg {
         left: &FieldElement,
         right: &FieldElement,
     ) -> Result<FieldElement, Error> {
-        use super::FIELD_BYTES;
-
         let lhs_ptr: usize = 0;
         let rhs_ptr: usize = lhs_ptr + FIELD_BYTES;
         let result_ptr: usize = rhs_ptr + FIELD_BYTES;
@@ -38,9 +36,6 @@ impl Pedersen for Barretenberg {
 
     #[allow(dead_code)]
     fn compress_many(&self, inputs: Vec<FieldElement>) -> Result<FieldElement, Error> {
-        use super::FIELD_BYTES;
-        use crate::barretenberg_structures::Assignments;
-
         let input_buf = Assignments::from(inputs).to_bytes();
         let input_ptr = self.allocate(&input_buf)?;
         let result_ptr: usize = 0;
@@ -52,9 +47,6 @@ impl Pedersen for Barretenberg {
     }
 
     fn encrypt(&self, inputs: Vec<FieldElement>) -> Result<(FieldElement, FieldElement), Error> {
-        use super::FIELD_BYTES;
-        use crate::barretenberg_structures::Assignments;
-
         let input_buf = Assignments::from(inputs).to_bytes();
         let input_ptr = self.allocate(&input_buf)?;
         let result_ptr: usize = 0;
