+++ src/barretenberg/scalar_mul.rs
